fix test
