+++ components/test_coprocessor/Cargo.toml
@@ -10,6 +10,7 @@ prost-codec = ["prost"]
 [dependencies]
 futures = "0.1"
 keys = { path = "../keys" }
+txn_types = { path = "../txn_types" }
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
 prost = { version = "0.5", optional = true }
 protobuf = "2"
