+++ components/test_coprocessor/src/store.rs
@@ -10,7 +10,7 @@ use test_storage::{SyncTestStorage, SyncTestStorageBuilder};
 use tidb_query::codec::{datum, table, Datum};
 use tidb_query::expr::EvalContext;
 use tikv::storage::{
-    txn::FixtureStore, Engine, Mutation, RocksEngine, SnapshotStore, TestEngineBuilder,
+    txn::FixtureStore, Engine, RocksEngine, SnapshotStore, TestEngineBuilder,
 };
 use tikv_util::collections::HashMap;
 use txn_types::{Key, TimeStamp};
