+++ components/test_raftstore/src/pd.rs
@@ -15,12 +15,13 @@ use kvproto::metapb::{self, Region};
 use kvproto::pdpb;
 use raft::eraftpb;
 
-use keys::{self, data_key, enc_end_key, enc_start_key, UnixSecs};
+use keys::{self, data_key, enc_end_key, enc_start_key};
 use pd_client::{Error, Key, PdClient, PdFuture, RegionInfo, RegionStat, Result};
 use tikv::raftstore::store::util::check_key_in_region;
 use tikv::raftstore::store::{INIT_EPOCH_CONF_VER, INIT_EPOCH_VER};
 use tikv_util::collections::{HashMap, HashMapEntry, HashSet};
 use tikv_util::timer::GLOBAL_TIMER_HANDLE;
+use tikv_util::time::UnixSecs;
 use tikv_util::{Either, HandyRwLock};
 
 use super::*;
