+++ components/test_storage/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 [dependencies]
 futures = "0.1"
 keys = { path = "../keys" }
+txn_types = { path = "../txn_types" }
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
 test_raftstore = { path = "../test_raftstore" }
 tikv = { path = "../../", default-features = false }
